Created AudioManager and Menus prefabs to be used in every level
Added a bunch of test sounds to the AudioManager script that we can swap out easily.
The AudioManager object & script will now control all audio in every scene, so I'll still need to remove that functionality from the GameManagers script and just make reference to the AudioManager script in it.
Menus prefab contains game settings, and will soon be put into a larger UI prefab to be used in every scene.